restructure user service
Restructuring user service to follow the new architecture pattern. This will allow for better organization/separtation for the different services in the backend.
