+++ backend/internal/app/user/user_service_test.go
@@ -1,4 +1,4 @@
-package app
+package user
 
 import (
 	"context"
